requirements.txt
googletrans==3.0.0
inflect==5.3.0
Pillow==9.0.0
pyttsx3==2.90
textblob==0.17.1
